refactor(agent): move BYOK probe execution out of routes (TURN-1)
probe_byok_model (Agent construction, ModelHTTPError taxonomy, message
assembly) moves to agents/byok_probe.py; routes/byok.py consumes the
neutral ProbeResult. photo_search.py drops the pydantic-ai Model import in
favor of a local structural Protocol. Route layer now has zero pydantic_ai
imports across byok/photo_search. 1755 unit tests green, ruff clean.

Written by opencode (async fleet, ds-flash-max)
