+++ apps/agent/src/animichi/interfaces/routes/photo_search.py
@@ -15,13 +15,12 @@ from __future__ import annotations
 
 from dataclasses import dataclass, field
 from datetime import UTC, datetime
-from typing import Annotated, Literal, cast
+from typing import Annotated, Literal, Protocol, cast
 
 import httpx
 from fastapi import APIRouter, Depends, Request
 from fastapi.responses import JSONResponse
 from pydantic import BaseModel, Field
-from pydantic_ai.models import Model
 
 from animichi.agents.base import get_default_model
 from animichi.agents.byok_models import ByokError, ByokModel, build_byok_model
@@ -89,11 +88,19 @@ class PhotoConfirmBody(BaseModel):
     candidates_shown: int = Field(ge=0)
 
 
+class _PlatformModel(Protocol):
+    """Structural model-provider type (satisfied by pydantic-ai models).
+
+    TURN-1 (#939): the route layer must not import framework model types; it
+    only carries the provider between domain helpers.
+    """
+
+
 @dataclass
 class PhotoSearchRuntime:
     """Per-app photo-search wiring, injectable for tests."""
 
-    platform_model: Model
+    platform_model: _PlatformModel
     catalog: CatalogClientProtocol
     quota: PhotoSearchQuota = field(
         default_factory=lambda: PhotoSearchQuota(clock=lambda: datetime.now(UTC))
@@ -190,7 +197,7 @@ async def _prepare_turn(
 
 def _recognize_call(
     runtime: PhotoSearchRuntime,
-    byok_model: Model | None,
+    byok_model: _PlatformModel | None,
     images: list[bytes],
     locale: str,
 ) -> RecognizeCall:
